purge the size_t s

+++ include/gloom/objectarray.h
@@ -4,9 +4,6 @@
 #include <libs>
 #include <Gloom/Dark2.h>
 
-// an objectarray wraps a grup
-// it makes iterating safer, more runtime resistent and better looking
-
 namespace gloom
 {
 	using Objects = ObjectArray;
@@ -17,45 +14,37 @@ namespace gloom
 	{
 	public:
 		Grup *const grup;
-
 		ObjectArray(Grup *);
-
-		inline int Type(size_t i) const
+		inline int Type(unsigned int i) const
 		{
 			return (int)((TypeDud *)_Get(i))->x;
 		}
-
-		inline size_t Size() const
+		inline unsigned int Size() const
 		{
 			return grup->mixed.size;
 		}
-
-		void *_Get(size_t i) const
+		void *_Get(unsigned int i) const
 		{
 			cassert(i < Size(), "objects bounds");
 			return grup->mixed.elements[i];
 		}
-
-		void *GetSafe(size_t i, int type) const
+		void *GetSafe(unsigned int i, int type) const
 		{
 			cassert(type == Type(i), "objects type x");
 			return _Get(i);
 		}
-
-		Grup *GetGrup(size_t i) const
+		Grup *GetGrup(unsigned int i) const
 		{
 			return (Grup *)GetSafe(i, GRUP);
 		}
-
-		Record *GetRecord(size_t i) const
+		Record *GetRecord(unsigned int i) const
 		{
 			return (Record *)GetSafe(i, RECORD);
 		}
-
 		template <class UnaryFunction>
 		void ForEach(int type, bool &stop, UnaryFunction f)
 		{
-			for (size_t i = 0; i < Size(); i++)
+			for (unsigned int i = 0; i < Size(); i++)
 			{
 				if (type == 0 || Type(i) == type)
 					f(*this, i);
