+++ src/gloom/objectarray.cxx
@@ -11,13 +11,13 @@ namespace gloom
 		
 		bool stop = false;
 		
-		objectArray.ForEach(RECORD, stop, [&](Objects &oa, size_t &i) {
+		objectArray.ForEach(RECORD, stop, [&](Objects &oa, unsigned int &i) {
 			Object object(oa.GetRecord(i));
 		});
 
 		using Objects = ObjectArray;
 
-		Objects(grup).ForEach(0, stop, [&](Objects &oa, size_t &i) {
+		Objects(grup).ForEach(0, stop, [&](Objects &oa, unsigned int &i) {
 			i += 4;
 			int type = oa.Type(i);
 			stop = true;
